Moving actors to underneath compiler
