+++ app/compiler/actors/AnalyzeSocketActor.java
@@ -9,35 +9,35 @@
  * This file is subject to the terms and conditions defined in
  * file 'LICENSE.txt', which is part of this source code package.
  */
-package actors;
+package compiler.actors;
 
 import akka.actor.ActorRef;
-import akka.actor.PoisonPill;
 import akka.actor.Props;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ObjectNode;
 import play.libs.Json;
 
 /**
- * <p>This class handles all request for generating VCs.</p>
+ * <p>This class handles all request for analyzing a RESOLVE file, which is
+ * simply populate and type check the file. This is mainly used to check to see
+ * if a theory file is valid or not.</p>
  *
  * @author Yu-Shan Sun
  * @version 1.0
  */
-public class VCSocketActor extends AbstractSocketActor {
+public class AnalyzeSocketActor extends AbstractSocketActor {
 
     // ===========================================================
     // Constructors
     // ===========================================================
 
-    /**c
-     * <p>This creates a new compiler job for generating VCs.</p>
+    /**
+     * <p>This creates a new compiler job for analyzing a file.</p>
      *
      * @param out Outgoing end of the stream.
      * @param job Name of the job to be executed.
      * @param project RESOLVE project folder to be used.
      */
-    public VCSocketActor(ActorRef out, String job, String project) {
+    public AnalyzeSocketActor(ActorRef out, String job, String project) {
         super(out, job, project);
     }
 
@@ -47,7 +47,7 @@ public class VCSocketActor extends AbstractSocketActor {
 
     public static Props props(ActorRef out, String job, String project) {
         // https://doc.akka.io/docs/akka/current//actors.html
-        return Props.create(VCSocketActor.class, () -> new VCSocketActor(out, job, project));
+        return Props.create(AnalyzeSocketActor.class, () -> new AnalyzeSocketActor(out, job, project));
     }
 
     /*@Override
@@ -56,23 +56,11 @@ public class VCSocketActor extends AbstractSocketActor {
             // Only deal with Strings
             if (message instanceof String) {
                 JsonNode request = Json.parse((String) message);
-
-                // Create a JSON Object informing we are starting the job
-                ObjectNode result = Json.newObject();
-                result.put("status", "info");
-                result.put(
-                        "msg",
-                        "Received request with the following parameters: "
-                                + myJob
-                                + " and "
-                                + myProject
-                                + ". Launching the RESOLVE compiler with the specified arguments.");
-
-                // Send the message through the websocket
-                myWebSocketOut.tell(result.toString(), self());
-
-                // Close the connection
-                self().tell(PoisonPill.getInstance(), self());
+                String[] args =
+                        { "-main", myWorkspacePath, "-webinterface", "Test.mt" };
+                TheoryAnalyzeInvoker invoker =
+                        new TheoryAnalyzeInvoker(args, myWebSocketOut);
+                invoker.executeJob(new HashMap<>());
             }
             else {
                 // Send an error message back to user and close
