+++ app/compiler/actors/ErrorSocketActor.java
@@ -9,7 +9,7 @@
  * This file is subject to the terms and conditions defined in
  * file 'LICENSE.txt', which is part of this source code package.
  */
-package actors;
+package compiler.actors;
 
 import akka.actor.ActorRef;
 import akka.actor.PoisonPill;
